Fix build errors
