+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import { AppProps } from 'next/app'
 import { RecoilRoot } from 'recoil'
-import { useOneSignal } from "@common/utils"
+import { useOneSignal } from '@common/useOneSignal'
 import '../styles/globals.css'
 
 export default function Ngosi({ Component, pageProps }: AppProps) {
